Compiles nearly cleanly under g++ 4. Commented out a use of long double for now. Added an explicit namespace to a template class friend declaration. Fixed several signed vs unsigned comparison warnings. Fixed some initialisation-order warnings (re #3)

+++ src/binary_reader.cpp
@@ -326,7 +326,7 @@ namespace protean {
         }
 
         // check magic number
-        if (header[0]!=binary_magic_number)
+        if (header[0]!=static_cast<int> (binary_magic_number))
         {
             boost::throw_exception(variant_error("Bad magic number, this looks like invalid binary data"));
         }
